requirements.txt
Click==7.0
Flask==0.12.4
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
pip==19.0.3
pyodbc==4.0.26
setuptools==40.8.0
Werkzeug==0.15.5
imutils==0.5.3
pytesseract==0.3.0
